Projeto pronto

+++ src/app/components/pages/moments/moments.component.ts
@@ -1,46 +1,82 @@
 import { Component, OnInit } from '@angular/core';
 import { Route, ActivatedRoute, Router } from '@angular/router';
 import { MomentService } from 'src/app/services/moment.service';
-
 import { Moment } from 'src/app/moment';
 import { environment } from 'src/environments/environment';
 import { MessagesService } from 'src/app/services/messages.service';
+import { FormGroup, FormControl, Validator, FormGroupDirective, Validators } from '@angular/forms';
+import { CommentService } from 'src/app/services/comment.service';
+import { Comment } from 'src/app/Comment';
 
 @Component({
   selector: 'app-moments',
   templateUrl: './moments.component.html',
   styleUrls: ['./moments.component.css']
 })
-export class MomentsComponent implements OnInit {
 
+export class MomentsComponent implements OnInit {
   moment?: Moment;
   baseApiUrl = environment.baseApiUrl;
+  commentForm!: FormGroup;
 
 
   constructor(
     private momentService: MomentService,
     private route: ActivatedRoute,
-    private messageService: MessagesService,
-    private router: Router
+    private messagesService: MessagesService,
+    private router: Router,
+    private commentService: CommentService
+
   ) { }
 
   ngOnInit(): void {
-
-    //id que está na URL
     const id = Number(this.route.snapshot.paramMap.get('id'));
-    this.momentService.getMoment(id).subscribe(item => this.moment = item.data);
+
+    this.momentService.getMoment(id).subscribe((item) => {
+      this.moment = item.data;
+    });
+
+    this.commentForm = new FormGroup({
+      text: new FormControl('', [Validators.required]),
+      username: new FormControl('', [Validators.required]),
+    });
   }
 
-  // Criar removeHandler
-    async removeHandler(id: number) {
-
-    await this.momentService.removeMoment(id).subscribe();
-
-    this.messageService.add('Momento removido com sucesso!');
-
-    this.router.navigate(['/']);
+  get text() {
+    return this.commentForm.get('text')!;
   }
 
+  get username() {
+    return this.commentForm.get('username')!;
+  }
 
+  async removeHandler(id: number) {
+    if (id) {
+      await this.momentService.removeMoment(id).subscribe();
 
+      this.messagesService.add(`Momento excluído com sucesso!`);
+
+      this.router.navigate(['/']);
+    }
+  }
+
+  async onSubmit(formDirective: FormGroupDirective) {
+    if (this.commentForm.invalid) {
+      return;
+    }
+
+    const data: Comment = this.commentForm.value;
+
+    data.momentId = Number(this.moment!.id);
+
+    await this.commentService
+      .createComment(data)
+      .subscribe((comment) => this.moment!.comments!.push(comment.data));
+
+    this.messagesService.add(`Comentário adicionado!`);
+
+    this.commentForm.reset();
+
+    formDirective.resetForm();
+  }
 }
